feat(rtp-core): move the RFC 4733 DTMF codec down from media-core
DtmfEvent and TelephoneEvent (the pure RFC 4733 telephone-event
encode/decode) lived in rvoip-media-core, but rtp-core's own UDP
transport already needed the same thing: transport/udp.rs had its own
inline copy of the 4-byte payload parsing in its receive loop, decoding
event/end-of-event/volume/duration by hand instead of reusing the type
that already existed one crate up. A consumer that only wants RFC 4733
without the rest of the media pipeline, such as something driving its
own mio reactor instead of RtpSession, had no way to get either version
without pulling in all of media-core.

Moved DtmfEvent and TelephoneEvent into rvoip-rtp-core::dtmf, since
media-core depends on rtp-core and not the other way around, so this
is a straight move down the dependency graph, not a new cycle. Added
serde as an rtp-core dependency (DtmfEvent/TelephoneEvent derive
Serialize/Deserialize; rtp-core didn't need serde before). Re-exported
both types from their original media-core path
(rvoip_media_core::codec::audio::dtmf) so existing consumers, including
dtmf_transmitter.rs, keep working without changing an import.

Wired transport/udp.rs's PT 101 receive path to call
TelephoneEvent::decode instead of its own inline parsing, so there's
one implementation instead of two that could drift apart. Behavior is
unchanged: a payload_type=101 packet still falls through to the
generic MediaReceived path when the payload is too short to decode,
same as before.

All rtp-core tests (271, including the 10 moved from media-core) and
all media-core tests (236 plus integration suites) pass.
